Separate actor and handler in files
